moved buildinginfo to model
