+++ app/src/main/java/org/cpen321/discovr/model/BuildingInformation.java
@@ -1,4 +1,4 @@
-package org.cpen321.discovr;
+package org.cpen321.discovr.model;
 
 /**
  * Created by jacqu on 11/9/2016.
